+++ src/Data/Poly.hs
@@ -28,7 +28,6 @@ module Data.Poly
 #if MIN_VERSION_semirings(0,4,2)
   -- * Polynomials over 'Field'
   , PolyOverField(..)
-  , gcdExt
   , PolyOverFractional
   , pattern PolyOverFractional
   , unPolyOverFractional
@@ -37,7 +36,7 @@ module Data.Poly
 
 import Data.Poly.Internal.Dense
 #if MIN_VERSION_semirings(0,4,2)
-import Data.Poly.Internal.Dense.Field (gcdExt)
+import Data.Poly.Internal.Dense.Field ()
 import Data.Poly.Internal.Dense.GcdDomain ()
 import Data.Poly.Internal.PolyOverField
 #endif