+++ src/Data/Poly/Sparse.hs
@@ -25,14 +25,10 @@ module Data.Poly.Sparse
   , subst
   , deriv
   , integral
-#if MIN_VERSION_semirings(0,4,2)
-  -- * Polynomials over 'Field'
-  , gcdExt
-#endif
   ) where
 
 import Data.Poly.Internal.Sparse
 #if MIN_VERSION_semirings(0,4,2)
-import Data.Poly.Internal.Sparse.Field (gcdExt)
+import Data.Poly.Internal.Sparse.Field ()
 import Data.Poly.Internal.Sparse.GcdDomain ()
 #endif