+++ changelog.md
@@ -1,6 +1,7 @@
 # 0.4.0.0
 
 * Implement orthogonal polynomials.
+* Decomission extended GCD, use `Data.Euclidean.gcdExt`.
 
 # 0.3.3.0